requirements.txt
streamlit
pandas
numpy
scikit-learn
matplotlib
python-docx
